+++ exts/cesium.omniverse/cesium/omniverse/ui/attributes/tileset_attributes_widget.py
@@ -61,6 +61,8 @@ class CesiumTilesetAttributesWidget(SchemaPropertiesWidget):
                 CustomLayoutProperty("cesium:smoothNormals")
             with CustomLayoutGroup("Georeference"):
                 CustomLayoutProperty("cesium:georeferenceBinding")
+            with CustomLayoutGroup("Point Clouds"):
+                CustomLayoutProperty("cesium:pointSize")
 
         return frame.apply(props)