+++ CHANGES.md
@@ -5,6 +5,7 @@
 * Added `pointSize` attribute to `CesiumTilesetPrim` for controlling the size of points.
 * Added read-only attribute `ecefToUsdTransform` to `CesiumGeoreferencePrim`. Previously this was stored in `/CesiumSession` which has since been removed.
 * Fixed crash when updating globe anchor when georeferencing is disabled.
+* Fixed point cloud styling.
 
 ### v0.20.0 - 2024-05-01